+++ CHANGELOG.md
@@ -9,6 +9,8 @@
 - Updated the config syntax inline with standard Nextflow style
 - Usage of tree structure for result value propagation and summary / accumulation
 - Configuration now represented in Nextflow Syntax Extension and V2 Parser
+- Provided an extension point so trace files can be processed post-run through a pipeline
+- Provided an CLI option to process execution trace files post-run
 
 # Version 1.1.0
 ## Bug Fixes: